Added support for multivalued "ns_set get|iget" results
In general, an ns_set can be multivalued, i.e., for one key, there
might be multiple values in the set. So far, it was necessary to
iterate over the values on the Tcl level. Now the option "-all" can be
provided to return a set of values.

   ns_set get  ?-all? setId key ?default?
   ns_set iget ?-all? setId key ?default?

In addition, the API function NsSetGetCmpDListAppend() was added on
the C-level. This function could also be used internally, when
e.g. multiple header-fields with the same key should be treated in
combination.

For providing easier support error messages for sub-commands, the
function Ns_ParseObjv2() was introduced, which has an additional
argument for handling leading words from the argument vector in error
messages.