New command "ns_ip"
The command "ns_ip" can be used to determine certain properties of IP
addresses. It accepts as input IPv4 or IPv6 addresses and checks, for
example, whether the address is valid, whether the IP is a public or
trusted IP address, or whether it is in the range of a subnet
specification in CIDR notation.

Essentially, the command makes internal functionality available at the
Tcl level.

The command provides the following subcommands:

   ns_ip match /CIDR/ /ipaddr/
   ns_ip properties /ipaddr/ -> dict containing type ipv4|ipv6, public 0|1 trusted 0|1
   ns_ip public /ipaddr/
   ns_ip trusted /ipaddr/
   ns_ip valid ?-type ipv4|ipv6? /ipaddr/

   ns_ip lookup ?-all? /ipaddr/

The command makes the command "ns_subnetmatch" obsolete, which is
still supported, but which will be deprecated in the future.