Do not prepend full location for redirects when running behind reverse proxy
When running behind one (or many) proxy servers, all attempts for
redirects to provide full location (with scheme and hostname) are
doomed, unless detailed (and trusted) client information is
provided. Therefore, in such cases, NaviServer does not try to prepend
the location.

This rule only kicks in when
1) connLocationProc is not configured, and
2) NaviServer vhosts are not enabled

In these cases, the user has already a provided detailed information,
so this will not be altered.