New subcommand: "ns_server ?-server /server/? hosts"
The subcommand of "ns_server" returns the registered hosts for this
server. These are the hostnames or IP addresses with optional ports
that are regarded as safe entries for received "host:" header
fields. The names are registered via the normal virtual host registration:

   ns_section ns/module/DRIVERNAME/servers {
      ns_param SERVER DOMAIN_NAME|IP?:PORT?
   }

The white-listed host entries are important, e.g., in cases, where the
provided host header field of a request is used for constructing
dynamically a fully qualified URL for a redirect. Without the
white-listed hosts, a spoofed entry could be used to hijack requests
to other servers via redirects.