ns_http: Added experimental support for Unix domain sockets
This change adds support for Unix domain sockets for HTTP client
operations.  Instead of running requests via HTTP/HTTPS, Unix domain
sockets are used for communicating to a server that is listening on a
Unix domain socket.  The interface is modeled after curl, using an
extra argument "-unix_socket". The host header information of the
request is taken from the URL.

For example, when a web server is listening on a Unix domain socket
named "/tmp/http.socket", ns_http can be used as the following:

   ns_http run -unix_socket /tmp/http.socket http://foo.org/

Current shortcomings:

- there is no locking provided, so i would suspect problems from
  concurrent requests to the same Unix domain socket (locking can be
  provided on the scripting level)

- There is no support for Unix domain sockets under Windows.