Added code to avoid repeated lines with the same log message
This implementation is based on thread local variables to avoid lock
operations. Therefore, we can only detect the same message coming from
the same thread - which is the most important case. Filtering happens
only, when thread local variables are supported (which is the case at
least with gcc/clang under Linux and MSC).