Avoid usage if deprecated C function mktemp()
Tthe pattern of creating a temporary file name based
on its non-occurrence in a directory followed by creating this file is
prone to a race condition and can provide security issues. Therefore,
the command is deprecated and will issue an entry to the log file
unless the option "-nocomplain" is used.

One should avoid this pattern and use "file tempfile" or functions
like "ns_uuid" to mitigate the issue.