Configuration file naming: make naming more straightforward
We do now allow the terms "map", "map-inherit" and "map-noinherit" for
mapping URLs to pools. Since "map" is accepted as a short form for
"map-inherit", the change is fully backward compatible.

In addition, the documentation of "mapspec" in the "ns_server" man page
provides more details about the relationship between the "mapspec" and
the URL space (as described in "ns_urlspace").