openacs-config.tcl: removed ${bindir} form module paths in sample configuration file
The changes are of the form:

   -        ns_param http ${bindir}/nssock
   +        ns_param http nssock

The leading "${bindir}" is not necessary, since in about every case,
the default for the ${bindir} (when it is omitted) is correct. When
leaving it, it opens a vector for misconfiguration, which can lead to
crashes, when e.g. NaviServer for Tcl8 and Tcl9 are both installed,
and the same configuration file is used and has a hard-coded path.