Prefer generic term "value" instead of "string" for placeholder arguments
The change effects the documentation and usage messages (in case of
errors) and therefore also the regression test.

Background: before NaviServer 5 the placeholder arguments were in many cases simply the repetition of the name of the parameter.
With NaviServer 5, the placeholders are replaced depending on the
converter, therefore, providing type information, including more
precise handle information, or enumerations

      [opt [option "-outlen [arg integer]"]] \
      [opt [option "-headers [arg setId]"]] \
      [opt [option "-samesite strict|lax|none"]] \

When no type information can be derived from the converter, the generic
term "string" was used. This change alters this to "value", because
sometimes, "string" might be misleading as well.

-     [opt [option "-filename [arg string]"]] \
-     [opt [option "-procs [arg string]"]] \
-     [opt [option "-pem [arg string]"]] \

+     [opt [option "-filename [arg value]"]] \
+     [opt [option "-procs [arg value]"]] \
+     [opt [option "-pem [arg value]"]] \