Bugfix: Don't use POLLPRI for detecting socket exceptions.
In the old version, when a sockcallback() requested callbacks on
errors (flag "e", via "ns_sockcallback" or "ns_connchan callback"),
this error callback was triggered (at least on macOS) also for certain
non-error conditions. The poll system called returned sometimes 0x0013
(meaning POLLIN | POLLPRI | POLLHUP), causing both triggering the "r"
and the "e" callback. The latter one was callback because the POLLPRI
mask matches. POLLPRI is actually meant to handle "out-of-band data on a
TCP socket" [1], which might occur in various situations, which are not
necessarily errors.

The new version uses now for the 'e' callback the POLLERR instead of
the POLLPRI flag, since POLLERR flags explicitly error conditions.

[1] https://man7.org/linux/man-pages/man2/poll.2.html