Feature request: provide error message to the user, when upstream request takes too long (connchan mode)
Before, we replied "504 Gateway Timeout" only in cases, when
experiencing a connection timeout to the upstream server.  In other
cases, the user saw just a blank screen.

With this change, we are now also able to send a user-visible reply,
when the reply takes too long. This change required some
refactoring. Before, revproxy sent the target headers immediately
after it was received by the proxy. Now we have to buffer the reply
headers and send it when we see actual data, or in different form when
an error is reported.