Reduced the required number of calls to the strlen() function
Earlier analysis showed that the function strlen() is one of the most
frequent C library calls of NaviServer, and that it uses a high number
of CPU cycles.  This change reduces the number of required calls to
the functions, mostly by passing Tcl_Objs instead of strings and using
Tcl_GetStringFromObj() instead of Tcl_GetString().