Adding support for streaming responses via ns_http
This commit adds streaming support for responses for requests issued
via ns_http.  This is necessary in cases, where the server sends
responses piece-wise and the responses must be processed before the
full request ends. Example application: using ns_http in a reverse
proxy implementation.

The corner points of the implementation are:

  - one can specify now for the "-outputchan" of ns_http a reverse
    proxy handle (no additional option needed).

  - since a reverse proxy has to return the headers from the backend
    server to achieve steaming results, a new response header callback
    was introduced, which is fired when the response headers from the
    target of the ns_http request are received.  This is in many
    respects similar to the "-donecallback". In the new callback, a
    developer can modify the headers and can send it also via connchan
    to the client.

Aside of the parameter for the new callback, the change is fully backward compatible.

Still open:
- more testing
- documentation updates
- reduce verbosity