Added revproxy module to main NaviServer repository
Previously, the revproxy module was a separate module
repository. Since the new version uses several new commands of
NaviServer 5, and the modules are not kept in multiple release
branches, the new version of the module was moved to the main
repository. The old module continues to work for the NaviServer 4.99
branch.

One of the interesting new features in the revproxy implementation is
using “ns_http” with a back-channel based on “ns_connchan”. This makes
it possible to use "ns_http" with streaming responses (streaming HTML,
streaming REST calls). "ns_connchan" is used for streaming the content
of the response of the backend server. This also works with persistent
connections for "ns_http" enabled.

One can use the new mode by specifying "ns_http+ns_connchan" in the
"backendconnection" configuration. The other modes continue to
work. So far, the default "backendconnection" is still "ns_connchan".

Another new feature is now the possibility to modify all request data
after the request was received from the client. This can be done via
the (optional) "validation_callback", which is the final request
before the request is forwarded to the backend. The
"validation_callback" receives now a "request" dict containing all
request data. The request dictionary contains the keys "headers",
"binary", and "content" or "contentfile". The value of "headers" is
the ns_set with the final request headers. The member "binary" is a
boolean value. When the content was spooled to a file, "contentfile"
contains the filename, otherwise `content` contains the string data of
the request body. The callback can modify all request data, and must
return the dict as result to continue.

The callback returns the same dict, which will be used later by the
revproxy implementation.  When the returned dict is empty, the request
is stopped. Therefore, this is a potential incompatibility with
previous releases of the revproxy module, where the result value of
the callback was always ignored.

Finally, the modules got a new documentation in Tcl doc style.

It is planned to use the revproxy implementation not only as a reverse
proxy, but as well for a classical "HTTP proxy" server, for which
NaviServer provides so far no full-featured implementation.

The version of the revproxy module was module bumped to 0.23.