Removed unused argument from prototypes of Ns_DriverSendProc and Ns_DriverSendFileProc
These two prototypes had since a long time a deactived timeoutPtr argument, which
was never used. The code had a TODO requesting a removal. The new major
release is a good time to remove such an argument.

Caveat: some network driver modules nscoap, nssyslogd, nstftpd, and nsudp
have to be recompiled for NaviServer 5. The modules are still source code
compatible for versions 4 and 5 of NaviServer.