Added explicit "-connecttimeout" to "ns_http run" and "ns_http queue"
Before, the connection timeout was the same as the general "-timeout"
value, which specifies the maximal time between socket events. While
later timeouts are dependent on the execution time of the request
(requiring e.g. database interactions), the connecttimeout is
independent of this.  Therefore, it can be set often to a smaller
value. The change is backward compatible, since the default value of
the "-connecttimeout" is the general "-timeout" value.