revproxy: use background operation for the spooling ns_http mode
This change lets delivery mode "ns_http+ns_connchan" run in the
background, particularly in the task thread. This change reduces the
number of required connection threads for revproxy deliveries, since
the connection thread does not have to wait for any result of "ns_http
run".

Currently, this background mode works only for "ns_http+ns_connchan"
but not for the pure "ns_http" delivery mode, since the latter needs
for the delivery still a connection thread for "ns_writer submit" and
"ns_writer submitfile".