Fix for non-delegitimized UNSPEC_TLS note issued by gcc-15
gcc-15 issues the following non-trivial note

   driver.c: In function 'SockSetServer':
   driver.c:5455:1: note: non-delegitimized UNSPEC UNSPEC_TLS (97) found in variable location
    5455 | SockSetServer(Sock *sockPtr)

It was caused by assigning the pointer to a thread‐local variable to a
structure which is reused by another thread. Since the thread-local
buffer may be reused or modified in subsequent calls, its pointer is
not safe to use for long-term storage.

In this commit, the "location" string is now duplicated (using
ns_strdup) so that the Sock structure holds its own copy of the
data. This prevents the warning and ensures the location string
remains valid regardless of further modifications to the thread-local
Tcl_DString.